requirements.txt
django==1.11.7
django-environ==0.4.4
djangorestframework==3.7.3
httpie==0.9.9
django-oauth-toolkit
django-cors-headers
gunicorn